HUGE refactor
- Moved `logLevel` to a persistenFlag, so no definition everywhere
- Seperated ucp_org.go into seperate ucp_auth_XX use cases
